requirements.txt
RPI.GPIO
numpy
pillow
transformers
torch
matplotlib
opencv-python